refactor: add conditions for delete payment method use case

+++ src/layers/application/use-cases/payment-method/delete-payment-method/delete-payment-method.use-case.ts
@@ -3,7 +3,8 @@ import {
 	DeletePaymentMethodDTO,
 	IDeletePaymentMethodUseCase,
 	NotFoundError,
-	IPayment
+	IPayment,
+	ForbiddenError
 } from "@/layers/application";
 
 export class DeletePaymentMethodUseCase implements IDeletePaymentMethodUseCase {
@@ -15,9 +16,15 @@ export class DeletePaymentMethodUseCase implements IDeletePaymentMethodUseCase {
 
 	async execute({ id }: DeletePaymentMethodDTO): Promise<string> {
 		const paymentMethodRepository = this.unitOfWorkRepository.getPaymentMethodRepository();
+		const subscriptionRepository = this.unitOfWorkRepository.getSubscriptionRepository();
 
 		const paymentMethod = await paymentMethodRepository.getPaymentMethodById(id);
 		if(!paymentMethod) throw new NotFoundError("Esse método de pagamento não existe");
+
+		const subscriptionActive = await subscriptionRepository.getActiveSubscriptionByUserId(paymentMethod.userId);
+		if(!subscriptionActive) throw new NotFoundError("O usuário não tem nenhuma assinatura ativa");
+		if(subscriptionActive.renewable) 
+			throw new ForbiddenError("Não possível excluir o método de pagamento pois existe uma assinatura ativa, cancele a assinatura para excluir o método de pagamento");
 
 		await this.unitOfWorkRepository.transaction(async () => {
 			const paymentMethod = await paymentMethodRepository.deletePaymentMethodById(id);
